General working algorithm for utility found
Ideally the score of the utility theory should be between 0 and 1.

Negative is the same result as 0 but positive need to be 1 or less to be scaled with other evaluators.

So start with base 1. Make a fraction that will result in base 1 like number of cards in hand. So if hand has 5 cards then base score is 5/5 = 1. Then the numerator can be manipulated by weighted values. e.g. mana is 3 then (5-3*.2)/5 = (5-.6)/5 = 4.4/5 = .88